kernel: Probe integrated modules immediately when they're loaded
Unfortunately, some devices spawn new devices upon probe, such as an i2c
master spawning new i2c slaves upon probe of the master. Since all drivers
are already loaded by the time integrated modules are allowed to probe,
this creates a very convoluted probe ordering issue where the i2c slaves
enumerated on a master are probed in the wrong order because they can all
probe immediately thanks to their drivers being loaded.

When a device is enumerated without having a driver registered for it, the
device isn't immediately probed and is instead autoprobed when a matching
driver is registered. This means that the order with which devices on a bus
are dynamically enumerated will respect the loading order of each driver,
provided that their driver isn't registered at the time they're probed.

Because integrated modules are loaded in one fell swoop with all probing
deferred until the very end, this means that the devices which are
enumerated on a bus at the time the bus' driver probes will all probe
immediately in the order with which they are enumerated on the bus,
violating the order in which the integrated modules are loaded. This can
lead to dependency bugs since devices can be probed in an unexpected order
with respect to the module load order.

Since the module load order must absolutely be upheld in order for
integrated modules to be reliable, integrated modules must be allowed to
probe immediately when they're loaded. Waiting until all drivers are loaded
before probing can expose dependency bugs in many drivers due to devices
getting probed in an unexpected order.

This still yields deterministic probe ordering because parallel module
loading and asynchronous probing are disabled for integrated modules. And
this still blocks processes which load modules from proceeding until all
the modules they loaded have had a chance to probe, so userspace can still
expect that an integrated module will be done probing after loading it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>